Correção dos seguintes bugs: - Crash na atividade de ordenar, ao ordenar rapidamente - Desorganização na atividade de ordenar, ao ordenar rapidamente - Popovers não sendo escondidas ao passar de fase, na atividade de montar ambiente - Objetos sumindo na atividade de montar ambiente
Melhoria no cálculo do frame da view de ordenação, para usar o espaço mais eficientemente.

+++ IntervencaoNarrativa/Categories/NSArray+Shuffling.m
@@ -27,7 +27,7 @@
     for (NSUInteger idx = 0; idx < count; ++idx)
 	{
         NSInteger remainingCount = count - idx;
-        NSInteger exchangeIndex = idx + arc4random_uniform(remainingCount);
+        NSInteger exchangeIndex = idx + arc4random_uniform((u_int32_t)remainingCount);
         [self exchangeObjectAtIndex:idx withObjectAtIndex:exchangeIndex];
     }
 }
